refactor: keep ETA sampling and estimation together
