+++ src/renderer/shared/format.ts
@@ -1,5 +1,5 @@
 import type { TaskSnapshot } from "../../task-model";
-import { isDeterminate } from "./determinate";
+import { estimateRemainingMillis } from "../../progress-estimation";
 import { getAmountParts } from "./amount-parts";
 
 const formatDurationSeconds = (seconds: number): string => {
@@ -27,42 +27,6 @@ const formatClockDurationSeconds = (seconds: number): string => {
   return hours > 0 ? `${`${hours}`.padStart(2, "0")}:${clock}` : clock;
 };
 
-/**
- * Estimates remaining time from the task's retained progress sample deque.
- *
- * Returns undefined until there are at least two samples with positive processed and time deltas.
- */
-const getSmoothedEtaMillis = (
-  task: TaskSnapshot & { readonly units: TaskSnapshot["units"] & { readonly total: number } },
-): number | undefined => {
-  const { processed, total } = task.units;
-  const remaining = total - processed;
-  if (processed <= 0 || remaining <= 0) {
-    return undefined;
-  }
-
-  const samples = task.progressSamples;
-  const lastSample = samples.at(-1);
-  if (lastSample === undefined) {
-    return undefined;
-  }
-
-  const firstSample = samples[0];
-  if (firstSample === undefined || firstSample === lastSample) {
-    return undefined;
-  }
-
-  // The store maintains this as a retained rolling deque, so the first and last samples represent
-  // recent throughput rather than lifetime average throughput.
-  const deltaProcessed = lastSample.processed - firstSample.processed;
-  const deltaMillis = lastSample.timestamp - firstSample.timestamp;
-  if (deltaProcessed <= 0 || deltaMillis <= 0) {
-    return undefined;
-  }
-
-  return Math.max(0, Math.floor((remaining * deltaMillis) / deltaProcessed));
-};
-
 export const formatElapsed = (task: TaskSnapshot, now: number): string => {
   const elapsedMillis = Math.max(0, (task.completedAt ?? now) - task.startedAt);
   return formatDurationSeconds(elapsedMillis / 1000);
@@ -76,17 +40,7 @@ const formatElapsedClock = (task: TaskSnapshot, now: number): string => {
 export const formatEta = (task: TaskSnapshot): string => formatEtaClock(task) ?? "";
 
 const formatEtaClock = (task: TaskSnapshot): string | undefined => {
-  if (task.status !== "running" || !isDeterminate(task)) {
-    return undefined;
-  }
-
-  const { processed, total } = task.units;
-  const remaining = total - processed;
-  if (processed <= 0 || remaining <= 0) {
-    return undefined;
-  }
-
-  const etaMillis = getSmoothedEtaMillis(task);
+  const etaMillis = estimateRemainingMillis(task);
   if (etaMillis === undefined) {
     return undefined;
   }
