+++ src/services/store/store.ts
@@ -1,3 +1,4 @@
+import { appendProgressSample } from "../../progress-estimation";
 import { Clock, Context, Effect, Layer, Option, Queue } from "effect";
 import type { Column } from "../../columns/types";
 import type { TaskId } from "../../task-model";
@@ -5,7 +6,6 @@ import type { TaskStore } from "./types";
 import type { TaskOperations } from "../task-operations";
 import { TaskId as makeTaskId, type TaskSnapshot } from "../../task-model";
 import {
-  appendProgressSample,
   createTaskSnapshot,
   finalizeTaskSnapshot,
   normalizeUnits,
