+++ src/services/store/task-state.ts
@@ -1,14 +1,11 @@
 import type { AddTaskOptions, UpdateTaskOptions } from "../../tasks/options";
-import type { TaskId, TaskProgressSample, TaskSnapshot } from "../../task-model";
+import type { TaskId, TaskSnapshot } from "../../task-model";
 
 interface TaskCounts {
   readonly succeeded: number;
   readonly failed: number;
   readonly total?: number;
 }
-
-const ETA_SAMPLE_WINDOW_MILLIS = 30_000;
-const ETA_SAMPLE_MAX_LENGTH = 1_000;
 
 const hasExplicitTotal = (options: Pick<AddTaskOptions | UpdateTaskOptions, "total">) =>
   Object.prototype.hasOwnProperty.call(options, "total");
@@ -51,38 +48,6 @@ const completedUnits = (units: TaskSnapshot["units"]): TaskSnapshot["units"] => 
         succeeded: units.succeeded + (units.total - units.processed),
       })
     : units;
-};
-
-/**
- * Returns a new progress sample deque with the latest processed count appended.
- *
- * Samples are retained for the ETA rolling window and capped by count so very chatty tasks do not
- * grow memory without bound.
- */
-export const appendProgressSample = (
-  samples: ReadonlyArray<TaskProgressSample> | undefined,
-  now: number,
-  processed: number,
-): ReadonlyArray<TaskProgressSample> => {
-  const previousSamples = samples ?? [];
-  const lastSample = previousSamples.at(-1);
-  if (lastSample?.processed === processed) {
-    return previousSamples;
-  }
-
-  // Keep one sample immediately before the rolling window when possible. That gives the ETA
-  // calculation a usable delta even just after old samples age out of the 30s window.
-  const windowStart = now - ETA_SAMPLE_WINDOW_MILLIS;
-  const appendedLength = previousSamples.length + 1;
-  let firstRetainedIndex = Math.max(0, appendedLength - ETA_SAMPLE_MAX_LENGTH);
-  while (
-    firstRetainedIndex + 1 < previousSamples.length &&
-    previousSamples[firstRetainedIndex + 1]!.timestamp < windowStart
-  ) {
-    firstRetainedIndex++;
-  }
-
-  return [...previousSamples.slice(firstRetainedIndex), { timestamp: now, processed }];
 };
 
 /** Applies mutable task fields; the mutation boundary records progress samples. */
